Change package to `io.spine.logging.*`
