+++ backends/logging-log4j2-backend/src/main/java/io/spine/logging/backend/log4j2/package-info.java
@@ -24,33 +24,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.google.common.flogger.backend.log4j2
-
-import io.spine.logging.backend.system.BackendFactory
-import io.kotest.matchers.collections.shouldHaveSize
-import io.kotest.matchers.nulls.shouldNotBeNull
-import io.kotest.matchers.types.shouldBeInstanceOf
-import java.util.*
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Test
-
 /**
- * Tests for [Log4j2BackendFactory].
- *
- * @see <a href="https://github.com/google/flogger/blob/cb9e836a897d36a78309ee8badf5cad4e6a2d3d8/log4j2/src/test/java/com/google/common/flogger/backend/log4j2/Log4j2BackendFactoryTest.java">
- *     Original Java code of Google Flogger</a>
+ * Contains implementation of `Log4j2` backend.
  */
-@DisplayName("`Log4j2BackendFactory` should")
-internal class Log4j2BackendFactorySpec {
+@CheckReturnValue
+package io.spine.logging.backend.log4j2;
 
-    @Test
-    fun `be loaded as a Java service`() {
-        val serviceLoader = ServiceLoader.load(BackendFactory::class.java)
-        val loadedServices = serviceLoader.toList()
-        loadedServices shouldHaveSize 1
-
-        val factory = loadedServices.first()
-        factory.shouldNotBeNull()
-        factory.shouldBeInstanceOf<Log4j2BackendFactory>()
-    }
-}
+import com.google.errorprone.annotations.CheckReturnValue;
