+++ backends/logging-log4j2-backend/src/main/resources/META-INF/services/io.spine.logging.backend.system.BackendFactory
@@ -1,1 +1,1 @@
-com.google.common.flogger.backend.log4j2.Log4j2BackendFactory
+io.spine.logging.backend.log4j2.Log4j2BackendFactory
