+++ backends/logging-log4j2-backend/src/test/kotlin/io/spine/logging/backend/log4j2/Log4j2LoggerBackendSpec.kt
@@ -24,11 +24,11 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.google.common.flogger.backend.log4j2
+package io.spine.logging.backend.log4j2
 
 import com.google.common.flogger.LogContext.Key
 import com.google.common.flogger.LogSite
-import com.google.common.flogger.backend.log4j2.given.MemoizingAppender
+import io.spine.logging.backend.log4j2.given.MemoizingAppender
 import com.google.common.flogger.parser.ParseException
 import com.google.common.flogger.repeatedKey
 import com.google.common.flogger.singleKey
