+++ backends/logging-log4j2-backend/src/test/kotlin/io/spine/logging/backend/log4j2/Log4j2ScopedLoggingSpec.kt
@@ -24,10 +24,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.google.common.flogger.backend.log4j2
+package io.spine.logging.backend.log4j2
 
 import com.google.common.flogger.LogContext.Key
-import com.google.common.flogger.backend.log4j2.given.MemoizingAppender
+import io.spine.logging.backend.log4j2.given.MemoizingAppender
 import com.google.common.flogger.context.ContextDataProvider
 import com.google.common.flogger.context.Tags
 import com.google.common.flogger.repeatedKey
