+++ backends/logging-log4j2-backend/src/test/kotlin/io/spine/logging/backend/log4j2/given/MemoizingAppender.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.google.common.flogger.backend.log4j2.given
+package io.spine.logging.backend.log4j2.given
 
 import org.apache.logging.log4j.core.LogEvent
 import org.apache.logging.log4j.core.appender.AbstractAppender
